refactoring: 제약 조건에 맞추어 리팩토링
